BlitzLol.RiotClient (blitz_lol v0.1.0)

Client for making requests against the Riot API

Types

region_type()

Specs

region_type() :: String.t()

A region defined by the Riot API. Can be one of br1, eun1, euw1, jp1, kr, la1, la2, na1, oc1, ru, tr1

routing_value_type()

Specs

routing_value_type() :: String.t()

A routing value defined by the Riot API. Can be one of americas, asia, or europe

Functions

fetch_match(routing_value, match_id)

Specs

fetch_match(routing_value_type(), String.t()) :: {:ok, map()} | {:error, any()}

Fetches a match from the Riot API

Example request:

fetch_match("americas", "NA1_4095008822")

Example response:

{:ok,
    %{
      "info" => %{
        "participants" => [%{...}]
      },
      "metadata" => %{}
    }}

fetch_matches(routing_value, puuid, params)

Specs

fetch_matches(routing_value_type(), String.t(), map()) ::
  {:ok, [String.t()]} | {:error, any()}

Fetches a list of matches for a given summoner puuid that matches all other criteria passed in through params

Example request:

fetch_matches("americas", "abc123", %{start: 0, count: 5})

Example response:

{:ok,
   ["NA1_4095008822", "NA1_4095014149", "NA1_4094938474", "NA1_4094905932", "NA1_4094903353"]}

fetch_summoner(region, summoner_name)

Specs

fetch_summoner(region_type(), String.t()) :: {:ok, map()} | {:error, any()}

Fetches a summoner from the Riot API

Example request:

fetch_summoner("na1", "mills")

Example response:

{:ok,
    %{
      "accountId" => "uiEWb5xUaBh6JrbLgc7DBM05MnKG6VXX2lDjicVlU10lGg",
      "id" => "uG3RHUMvuUPzq7fuJF8IdslZt8gzxslNxwgX91GdLUHLtxk",
      "name" => "Mills",
      "profileIconId" => 940,
      "puuid" =>
        "WcHwXcrQJbyqgnsXGcYlvTrEPhAcB2e11pofFCioEcKEwEkRMaMD9qrYIkzz5-MVx4q6tZZsjBIBxA",
      "revisionDate" => 1_636_252_885_672,
      "summonerLevel" => 245
    }}