BlitzLol.Summoners (blitz_lol v0.1.0)

Context module for interacting with summoners

Types

region_type()

Specs

region_type() :: String.t()

A region defined by the Riot API. Can be one of br1, eun1, euw1, jp1, kr, la1, la2, na1, oc1, ru, tr1

Functions

track_recent_summoners_by_summoner_name(region, summoner_name)

Specs

track_recent_summoners_by_summoner_name(region_type(), String.t()) ::
  {:ok, [String.t()]} | {:error, any()}

Returns a list of summoners that a summoner has played with in their last 5 matches.

The summoners will then be monitored for the next hour, and will check for a newly completed match every minute. If a new match is detected, then the summoner will be logged out along with the match they completed.

Example request:

track_recent_summoners_by_summoner_name("na1", "mills")

Example response:

{:ok,
  ["Tom The Killer", "Jeam kai", "styrofoamcond0m", "ZubatUsedLowKick", "Krhaal",
    "Snytz", "Kamstuh", "Nef", "Rogue Armadillo", "greNinjaSquid", "esterxk",
    "the derpy guy", "tamish3", "BigSid14", "Eurynnome", "Madarame Ikkaku",
    "Foxiest Boxes", "ThyneBowlCut", "LOAFman2020", "EXP3RIENCE", "Lóng gone",
    "JoshMcDooDoo", "Slain", "Flzcko", "1HPNotEvenClose", "Picheese", "nex991",
    "HerbalHelper25", "Kosher Dubstep", "StanislawLeafrak", "EmErEldSquEd",
    "TheFilipinosMom", "Andpv", "SauceZzz", "Chef Pasquale", "NO69U",
    "ExpIosìon", "Raghadast", "Gugulaka", "PANDAMAN3659", "Deleted282"]}